requirements.txt
bcrypt==4.3.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
et_xmlfile==2.0.0
eventlet==0.39.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-PyMongo==2.3.0
Flask-SocketIO==5.5.1
geographiclib==2.0
geopy==2.4.1
googlemaps==4.10.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.2
openpyxl==3.1.2
packaging==24.2
pandas==2.2.3
pycparser==2.22
PyJWT==2.10.1
pymongo==4.11.3
python-dateutil==2.9.0.post0
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2025.1
requests==2.32.3
simple-websocket==1.1.0
six==1.17.0
tzdata==2025.1
urllib3==2.3.0
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
xlrd==2.0.1
